Add user profile push preference UI to the Demo App
